video-selection/requirements.txt
fastapi>=0.80.0
uvicorn[standard]>=0.18.0
google-api-python-client>=2.0.0
pydantic
python-dotenv # For local development to load .env file